+++ test/TestAsset/ContainsUseFunction.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ZendTest\File\TestAsset;
+
+use function strlen;
+
+class ContainsUseFunction
+{
+
+}